Fixed rendering of tweets with emojis
refs https://github.com/TryGhost/Team/issues/1001

We cannot use the String methods like slice as they operate on UTF values,
but twitter uses Unicode to support emojies. Instead we split the string
into an array of codepoints, which will count an emoji as one item. With
this we can correctly use the `length` property and slice methods, we
just need to remember to call join to transform back to a string.

We also must order the entities that we're operating on so that we do
not concatenate the content string out of order.

Twitter sometimes includes newline characters as part of the url object
- so we have to transform these to <br> tags inline, rather than
  assuming they will be part of the plain text.