Added `ExternalControlPlugin` to let consumer apps handle tasks via a basic API (#420)
refs https://github.com/TryGhost/Team/issues/2286

We want to enable consumer apps to streamline the cursor handling between other inputs and the editor canvas, the primary use-case being a title input that should move the cursor to the editor as if it was part of the editor canvas itself.

By exposing a simplified API for typical actions we allow consumer apps to use it without knowing the intricacies of the Lexical API. It also means when using Koenig as an externally loaded library, the consumer app doesn't need to pull in any of the Lexical utility functions to it's own code.

- adds a `registerAPI` prop to the `<KoenigEditor>` component that is called with an object containing a basic API with the necessary features needed for title input->editor behaviour
- updated demo app to use the new plugin in it's title component
  - <kbd>Tab</kbd> moves cursor to the editor
  - <kbd>Enter</kbd> moves cursor to the editor, inserting a blank paragraph at the top of the document if the editor already has content
  - <kbd>Arrow Down/Right</kbd> moves cursor to the editor only when the cursor is at the end of the title input, maintaining default browser/OS behaviour up until that point