Fixed creation of cards via slash menu at the beginning of the document
closes https://github.com/TryGhost/Team/issues/2831

When inserting a card via the slash menu we first delete the contents of the current paragraph containing the command, however that causes issues when the paragraph is the first node in the document because it's `collapseAtStart()` method will completely remove the node when it's emptied. That lead to errors when the paragraph was followed by a card or incorrect ordering when followed by text content.

- adjusted slash menu insert to create a new paragraph and delete the old one to avoid any inconsistency between start and middle of document usage