🎨 Improved default settings panel position on mobile
no issue

There is no reason to display the settings panel in the middle of the screen when editing a card. This change makes sure we try to position the settings panel at the bottom of the screen instead. That makes it easier to edit the card content and move the settings panel in view.