🐛 Fixed cursor position reset when a card with a focused nested editor is removed (#850)
closes https://github.com/TryGhost/Product/issues/3430

Nested editors use a plugin that handles resetting focus in the parent editor when it loses its own focus so we don't lose the overall card selection. That behaviour was also triggering when the card is removed such as when pressing Cmd+Enter to exit out of an empty card triggering the auto-remove-empty behaviour resulting in a secondary incorrect reset of the cursor position.

- the use of `$getSelection()` outside of a `$editor.update()` closure meant the logic was incorrect as it was comparing against the nested editor's selection rather than the parent editor's selection
- moved `$getSelection()` inside the `$editor.update()` call so we aren't triggering the blur behaviour when the parent editor already has a selection