Switched to using mutationListener in TKPlugin (#1119)
no issue

Using a Lexical `mutationListener` is more performant because we only run code when a TKNode is added/modified/removed rather than on every keystroke. It also lets us simplify the code a bit because we can deal with nodes one-by-one in the context object instead of fetching all of them in TKPlugin and manipulating into a map to then store in the context.

- updated `TKContext` to expose methods for managing TK nodes
  - `addEditorTkNode()` / `removeEditorTkNode()` to be called from within the mutation listener
  - `removeEditor()` to be called from an effect run when an editor is removed to perform cleanup (mutation listener doesn't fire in this case)
  - these methods modify a map that is stored in a ref rather than updating state directly because the methods will be called many times when initiating an editor containing TK nodes and React doesn't handle that well if state is updated each time. After modifying the ref'd map they call a debounced update method that generates the tkNode state used for rendering TK indicators
- updated `TKContext` to use a simpler map model
  - by keying by editor and tkNode keys it makes adding/removing nodes one-by-one a lot easier
- updated `TKPlugin` to use the new context methods