Fixed "no results" state flashing in some situations (#1266)
ref https://linear.app/tryghost/issue/MOM-160

In Ghost's Ember host app, promises ("tasks") can be cancelled which will occur when the searchLinks function is called in quick succession. That was resulting in a flash of "no results" because promises in React/plain JS don't get cancelled so earlier search promises could end up overwriting the results array whilst a later search was still completing.

- added handling for receiving `undefined` from the `searchLinks` function which indicates a cancelled search
- fixed constant re-creation of search options object prop which caused problems in the hook when it's properties were used in dependency arrays