Added Captcha to the signup form using Portal's injection
ref EMAIL-105

Here the promise then chain needed to be converted to an async
function to make it easier to reason about the Captcha, which is
executed if the labs flag is enabled.

This code still needs additional tests, but we're at a point where the
functionality works, and existing tests pass.