Updated ghost-cli tests to update from latest v5 version (#24659)
The `ghost-cli` tests on the v5 branch are failing because they are
trying to upgrade from the "Latest release" (which is now v6) to the
current version — this is effectively now a downgrade on the `5.x`
branch, so the tests fail.

This changes it to attempt an upgrade from the latest public v5 release
to the current commit.