Manually importing TailwindCSS classes
Importing everything from Tailwind manually instead of adding it via @import 'tailwindcss'

The reason is because Tailwind uses CSS layers and there are some legacy CSS
in the Admin (mostly Spirit) with the same classnames as in TailwindCSS which
don't use any CSS layering. TWCSS v4 layering sets the specificity of classnames
lower than non-layered classnames so legacy classes would override TWCSS classes
even if the CSS classname specificity would be higher. For example:

```
.flex {display: flex};
```

in the Admin's `globals.css` overrides

```
@layer utilities { .shade .grid {display: grid}}
```

If we import tailwindcss files directly we can omit the use of layers. This is
not ideal but works with the current system.

Note that another option could have been to add layer to legacy classes but
that would have resulted in a much larger and riskier change.